add config option to hide development tools
