+++ backend/src/test/java/eu/clarin/switchboard/core/DataStoreTest.java
@@ -1,6 +1,6 @@
 package eu.clarin.switchboard.core;
 
-import eu.clarin.switchboard.app.Config;
+import eu.clarin.switchboard.app.config.DataStoreConfig;
 import eu.clarin.switchboard.core.xc.StoragePolicyException;
 import org.junit.Before;
 import org.junit.Test;
@@ -30,7 +30,7 @@ public class DataStoreTest {
         String maxLifetimeUnit = "seconds";
         String cleanupPeriod = "1";
         String cleanupPeriodUnit = "seconds";
-        Config.SwitchboardConfig.DataStoreConfig dataStoreConfig = new Config.SwitchboardConfig.DataStoreConfig(
+        DataStoreConfig dataStoreConfig = new DataStoreConfig(
                 dataStoreRoot.toString(), false, maxSize, maxLifetime, maxLifetimeUnit, cleanupPeriod, cleanupPeriodUnit);
         StoragePolicy storagePolicy = new DefaultStoragePolicy(dataStoreConfig);
         dataStore = new DataStore(dataStoreRoot, storagePolicy);
